:alien: Retry the completion call to Camunda three times
Camunda performs optimistic locking, which results in rolled back
transactions on conflict (which happens when parallel tasks arrive at
the same time, for example. The recommended approach by Camunda is to
retry the task, which effectively staggers the calls/updates.

Retries are only performed on internal server errors (HTTP 500) from
Camunda. If the error is of persistent nature, it will still bubble
up.