:goal_net: Ensure that failure to complete the task eventually lead to task failure
... after all retries have been exhausted.

This prevents tasks from having their lock expire and being picked
up again by BPTL when there's a problem with the variables, causing
Camunda to fail at advancing to the next process task.

Camunda uses thread-borrowing, which means that the process execution
happens in the thread serving the task-complete call. The process
execution continues to the next waiting points. However, sometimes
this fails because while preparing the next waiting point, an
expected variable in the process definition is not present. Camunda
then throws a 500 error on the completion. The task is not marked
as completed, and after lock expiry retried. This is an infinite
loop, cluttering the dashboard and logs. By allowing three
retries and failing the task after the third retry, the loop is broken.