:sparkles: Implemented Camunda long polling
Added celery-once to prevent multiple workers from opening long-poll
connections to Camunda.

When the task exits (both succesfully or on exception), it re-schedules
itself to open a new connection as soon as possible. In the event that
this would fail, beat also makes sure to try to fetch and lock tasks
every 10 seconds, which is fine because of celery-once managing a
shared lock.

Currently, the long-poll connection is kept open for a maximum of 10min,
if all is well, we can increase this to the maximum of 30 minutes.

The end result is that tasks should be picked up pretty much as they
are scheduled in Camunda, leading to a more fluent user-experience.