:whale: Put nginx in front to handle Content-Transfer: chunked
the wsgi spec does not support chunked transfer encoding, and runserver/uwsgi/django
break on that. Xential sends the callback body with this transfer encoding, causing
django not to be able to read the data. For deployment, we will put nginx in
front of uwsgi, since the Traefik ingress does not buffer the request body, but
nginx does. This will mean Traefik -> nginx -> uwsgi in deployment.